requirements.txt
flask>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=0.21.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0